StableDiffusion/requirements.txt
# --- Core Libraries ---
flask==3.0.2
gunicorn==21.2.0
numpy==1.26.4
pillow==11.0.0
tqdm==4.67.1

# --- ML & DL ---
pytorch_lightning==2.2.4
torchmetrics==1.3.2
lightning-utilities==0.10.1
transformers==4.45.2
tokenizers==0.20.3
safetensors==0.5.3

# --- Transformers Dependencies ---
huggingface-hub==0.33.1
filelock==3.13.1
requests==2.32.3
regex==2024.11.6
PyYAML==6.0.2
packaging==24.1
fsspec==2024.2.0
typing-extensions==4.9.0

# --- Flask Dependencies ---
Werkzeug==3.1.3
Jinja2==3.1.3
itsdangerous==2.2.0
click==8.1.7
blinker==1.9.0

# --- Requests Dependencies ---
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
urllib3==2.2.3

# --- Optional (for colored tqdm on Windows) ---
colorama==0.4.6 ; sys_platform == 'win32'